Lmdb-zero: Split into multiple files.
There's a certain awkwardness here and there due to the way certain
types are intertwined (Environment + Database; Transaction + Cursor),
but it was less painful than I expected.

It would be nice though if Rust had a way to control visibility of
inherent methods on a type though; the free functions aren't that nice,
and the alternative is typing everything twice by using an internal
extension trait.