Document the GCC issue that can be encountered with `DUPSORT`.
For those curious, with optimisations enabled, GCC performs the following
translation in `mdc.c`:

```
6590							mp->mp_ptrs[i] = fp->mp_ptrs[i] + offset;
   0x000055555569c5d8 <mdb_cursor_put+3816>:	mov    rax,QWORD PTR [rsp+0x20]
   0x000055555569c5dd <mdb_cursor_put+3821>:	mov    r9,rsi
   0x000055555569c5e0 <mdb_cursor_put+3824>:	add    rsi,0x1
   0x000055555569c5e4 <mdb_cursor_put+3828>:	shl    r9,0x4
   0x000055555569c5e8 <mdb_cursor_put+3832>:	cmp    r8d,esi
=> 0x000055555569c5eb <mdb_cursor_put+3835>:	movdqa xmm1,XMMWORD PTR [rax+r9*1]
   0x000055555569c5f1 <mdb_cursor_put+3841>:	mov    rax,QWORD PTR [rsp+0x10]
   0x000055555569c5f6 <mdb_cursor_put+3846>:	paddw  xmm1,xmm0
   0x000055555569c5fa <mdb_cursor_put+3850>:	movups XMMWORD PTR [rax+r9*1],xmm1
```

The issue is that `fp` is not 16-byte aligned, and thus `fp->mp_ptrs` is
not either, though it happens to be exactly 16 bytes after the start of
the structure. (The structure itself does not have an ABI alignment of
16 bytes, but rather only 8. There's probably an argument to be made
that LMDB is wrong by not aligning the structure even that much, but
even if it were so aligned, the vectorisation still seems wrong on
multiple counts.)

I'm honestly not sure whose bug this is; I'm going to think on it some
more and then hopefully remember to report it somewhere.