Problem: passing transaction and accessor around is not trivial
It's (nearly?) impossible to use accessors in scenarios that
fall out of commonly suggested pattern of:

```rust
{
  let accessor = txn.access();

  // work with the database
}

txn.commit();
```

Particularly, if the transaction is done in a few places
across the codebase, passing accessor throughout poses
a challenge.

Solution: implement Drop trait for ConstAccessor so that if it
gets dropped, a new one can be easily obtained.