Switch most internal references to `Supercow`s.
This is the first step in improving the ergonomics wrt lifetimes,
allowing use of owned values or smart pointers where preferable.

The acessor family is deliberately _not_ included in this change; it
will continue using plain references. Instead, the accessors will be
made more flexible by adopting
https://github.com/AltSysrq/lmdb-zero/pull/5 to simply allow dropping
and reobtaining the accessor.

There's still a decent bit to do:

- It isn't yet possible to use all ownership modes with all APIs. For
  example, one cannot construct a `Cursor` with shared ownership of the
  `ConstTransaction`.

- We need some additional APIs to work with certain cases, such as the
  one-db-owning-the-env model.

- Improve the way accessors work as mentioned above.

- Documentation.